+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+## 0.0.2
 
 * First release.
+
+## 0.0.3
+
+* Provider more information to pointer cancel events.
+* New `enabled` parameter.